Added Agent Movements, collisions, band aid fix for bug
Humans spawn and move randomly. Zombies chase humans and turn humans to zombies when colliding. The openGL memory leak caused by multiple spritebatches is still unfixed so I have created a work so that I can keep working on the game as it was only intended to be an unoptimized initial project